Update Deque.java
Hi! Thank you for the solutions!
The earlier code was not working on my system and showing compilation error because it could not interpret:

1) import edu.princeton.cs.algs4.StdOut;
The code is working without edu.princeton.cs.algs.StdOut because I don't have this file.

2) StdOut
System.out is the by-default function in my IntelliJ java. That's why i replaced all StdOut with System.out in my branch. 

I think this new code works perfectly for all computers and will be helpful to solve the assignment.
Thank you! Looking forward to a reply.